V.1.1

+++ CuentaBanco.java
@@ -1,13 +1,18 @@
 package Proyecto_CODE;
 
+/**
+ * Clase que define objetos de tipo CuentaBanco
+ * @author Alejandro Vidal
+ * @version  1.0
+ */
 public class CuentaBanco {
     private String DNI;
     private int  NumCuenta;
-    private int saldo = 0;
+    private double saldo = 0;
     private String propietario;
     //Constructores
 
-    public CuentaBanco(String DNI, int numCuenta, int saldo, String propietario) {
+    public CuentaBanco(String DNI, int numCuenta, double saldo, String propietario) {
         if(comprobar(DNI)){
             this.DNI = DNI;
         }
@@ -26,10 +31,18 @@ public class CuentaBanco {
     }
     //Metodos Get y Set
 
+    /**
+     * Metodo que permite recuperar el DNI
+     * @return String DNI
+     */
     public String getDNI() {
         return DNI;
     }
 
+    /**
+     * Metodo que permite asignar un DNI  mientras sea valido si no devuelve un mensaje
+     * @param DNI String al que se le asigna el valor pasado como parametro
+     */
     public void setDNI(String DNI) {
         if(comprobar(DNI)){
             this.DNI = DNI;
@@ -39,26 +52,50 @@ public class CuentaBanco {
         }
     }
 
+    /**
+     * Metodo que permite recurerar el valor de NumCuenta
+     * @return int numCuenta
+     */
     public int getNumCuenta() {
         return NumCuenta;
     }
 
+    /**
+     * Metodo que permite asignar un valor pasado como parametro a NumCuenta
+     * @param numCuenta int al que se asigna el valor
+     */
     public void setNumCuenta(int numCuenta) {
         NumCuenta = numCuenta;
     }
 
-    public int getSaldo() {
+    /**
+     * Metodo que permite recurerar el valor de saldo
+     * @return double saldo
+     */
+    public double getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(int saldo) {
+    /**
+     * Metodo que permiteasignar un valor pasado como parametroa saldo
+     * @param saldo double al que se asigna valor
+     */
+    public void setSaldo(double saldo) {
         this.saldo = saldo;
     }
 
+    /**
+     * Metodo que permite recuperar el valor almacenado en propietario
+     * @return String propietario
+     */
     public String getPropietario() {
         return propietario;
     }
 
+    /**
+     * Metodo que permite asignar un valor a propietario pasandolo como parametro
+     * @param propietario  String al que se le asigna valor
+     */
     public void setPropietario(String propietario) {
         this.propietario = propietario;
     }
@@ -72,6 +109,11 @@ public class CuentaBanco {
     }
     //Metodos de la clase
 
+    /**
+     * Metodo que verifica DNI
+     * @param cadena String que contiene un DNI
+     * @return devuelve true cuando el DNI cumple las restricciones y false cuando no
+     */
     public boolean comprobar(String cadena){
         String DNI = cadena.toUpperCase();
         int numero;
